# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClubHouse single-college club management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get own profile
      security: [{ clerk: [] }]
      responses:
        "200":
          description: Profile returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update own profile
      security: [{ clerk: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by name or email
      security: [{ clerk: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 8
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get any user's public profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found

  # ── COLLEGES ──────────────────────────────────────────────────────────────────
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [clubs]
      summary: Request a presigned upload URL
      security: [{ clerk: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageUploadInput"
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUploadResponse"
...
  /reports:
    post:
      operationId: createReport
      tags: [reports]
      summary: Report a project, club, event or profile
      security: [{ clerk: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "409":
          description: Already reported
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports by status (admin)
      security: [{ clerk: [] }]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, dismissed]
            default: open
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
  /reports/{reportId}:
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Resolve or dismiss a report (admin)
      security: [{ clerk: [] }]
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [resolved, dismissed] }
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
components:
  securitySchemes:
    clerk:
      type: http
      scheme: bearer
  schemas:
...
    SocialLinks:
      type: object
      properties:
        linkedin:   { type: ["string", "null"] }
        github:     { type: ["string", "null"] }
        instagram:  { type: ["string", "null"] }
        facebook:   { type: ["string", "null"] }
        reddit:     { type: ["string", "null"] }
        whatsapp:   { type: ["string", "null"] }
    PortfolioProject:
      type: object
      required: [id, title, url]
      properties:
        id:          { type: integer }
        title:       { type: string }
        url:         { type: string }
        description: { type: ["string", "null"] }
    UserProfile:
      type: object
      required: [clerkId, name, email, age, course, semester, college, pronouns, createdAt]
      properties:
        clerkId:         { type: string }
        name:            { type: string }
        email:           { type: string }
        age:             { type: integer }
        course:          { type: string }
        semester:        { type: integer }
        college:         { type: string }
        pronouns:        { type: string }
        bio:             { type: ["string", "null"] }
        avatarUrl:       { type: ["string", "null"] }
        portfolioProjects:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioProject"
        socials:
          $ref: "#/components/schemas/SocialLinks"
        role:            { type: string }
        createdAt:       { type: string, format: date-time }
    UserProfileUpdate:
      type: object
      properties:
        name:       { type: string }
        age:        { type: integer }
        course:     { type: string }
        semester:   { type: integer }
        college:    { type: string }
        pronouns:   { type: string }
        bio:        { type: ["string", "null"] }
        avatarUrl:  { type: ["string", "null"] }
        portfolioProjects:
          type: array
          items:
            type: object
            required: [title, url]
            properties:
              title:       { type: string }
              url:         { type: string }
              description: { type: ["string", "null"] }
        socials:
          $ref: "#/components/schemas/SocialLinks"

    # ── COLLEGE ──────────────────────────────────────────────────────────────────
...
    StorageUploadInput:
      type: object
      required: [name, size, contentType]
      properties:
        name:        { type: string }
        size:         { type: integer, minimum: 1 }
        contentType: { type: string }
        visibility:  { type: string, enum: [public, private], default: public }
    StorageUploadResponse:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL:  { type: string, format: uri }
        objectPath: { type: string }
        metadata:
          type: object
          required: [name, size, contentType]
          properties:
            name:        { type: string }
            size:        { type: integer }
            contentType: { type: string }

    # ── PROJECT ──────────────────────────────────────────────────────────────────
...
    Report:
      type: object
      required: [id, reporterId, targetType, targetId, reason, status, createdAt]
      properties:
        id:          { type: integer }
        reporterId:  { type: string }
        targetType:  { type: string, enum: [project, club, event, profile] }
        targetId:    { type: string }
        reason:      { type: string }
        description: { type: ["string", "null"] }
        status:      { type: string, enum: [open, resolved, dismissed] }
        createdAt:   { type: string, format: date-time }
        resolvedAt:  { type: ["string", "null"], format: date-time }
        resolvedBy:  { type: ["string", "null"] }
    ReportInput:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType:  { type: string, enum: [project, club, event, profile] }
        targetId:    { type: string }
        reason:      { type: string }
        description: { type: ["string", "null"] }
...